refactored topic tag
